retui: App inherits Pane
move TerminalWidget to widgets
and make App inherit Pane
half of functionality was duplicated
